requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
tavily-python==0.7.5
python-dotenv==1.0.1
requests==2.31.0
openai==1.3.0
jinja2==3.1.2
pytest==7.4.3
httpx==0.25.1